+++ data/workflow/easysearch.sh
@@ -74,5 +74,7 @@ if [ -n "${REMOVE_TMP}" ]; then
     rm -rf "${TMP_PATH}/search_tmp"
     rm -f "${TMP_PATH}/easysearch.sh"
     rm -rf "${TMP_PATH}"
-    rm -f "${LATEST}"
+    if [ "$(basename "${TMP_PATH}")" = "$(readlink "${LATEST}")" ]; then
+        rm -f "${LATEST}"
+    fi
 fi
