+++ data/workflow/easytaxonomy.sh
@@ -83,5 +83,7 @@ if [ -n "${REMOVE_TMP}" ]; then
     rm -rf "${TMP_PATH}/taxonomy_tmp"
     rm -f "${TMP_PATH}/easytaxonomy.sh"
     rm -rf "${TMP_PATH}"
-    rm -f "${LATEST}"
+    if [ "$(basename "${TMP_PATH}")" = "$(readlink "${LATEST}")" ]; then
+        rm -f "${LATEST}"
+    fi
 fi
